Repozytorja

+++ Bank.backend/src/main/java/pl/bj/bank/repository/KlientRepository.java
@@ -3,5 +3,5 @@ package pl.bj.bank.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.bj.bank.model.Klient;
 
-public interface KlientRepository extends JpaRepository<Klient, Long> {
+public interface KlientRepository extends JpaRepository<Klient, Integer> {
 }
